requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
python-dateutil==2.8.2
Pillow==10.0.1
newspaper3k==0.2.8
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0
PyGithub==1.59.1